# backend/docs/swagger.yaml
basePath: /
definitions:
...
  httpx.Error:
    properties:
      detail:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: LHCHub home-services marketplace API (Go backend). Clients post jobs,
    vetted providers apply, and payment is held in escrow.
  termsOfService: http://swagger.io/terms/
  title: LHCHub API
  version: "1.0"
paths:
  /api/auth/2fa/disable/:
    post:
      consumes:
      - application/json
      description: Disables 2FA for the current user.
      parameters:
      - description: '{password}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - auth
  /api/auth/2fa/enable/:
    post:
      consumes:
      - application/json
      description: Generates a TOTP secret and backup codes for the current user.
      parameters:
      - description: '{password}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Enable 2FA
      tags:
      - auth
  /api/auth/2fa/login-verify/:
    post:
      consumes:
      - application/json
      description: Verifies a TOTP or backup code against a preAuthToken and issues
        JWT tokens.
      parameters:
      - description: '{preAuthToken, code}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Complete 2FA login
      tags:
      - auth
  /api/auth/2fa/verify/:
    post:
      consumes:
      - application/json
      description: Confirms a TOTP code and activates 2FA for the current user.
      parameters:
      - description: '{code}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Verify 2FA
      tags:
      - auth
...
  /api/auth/otp/send/:
    post:
      consumes:
      - application/json
      description: Emails a one-time code for verification.
      parameters:
      - description: '{email, type}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Send email OTP
      tags:
      - auth
  /api/auth/otp/verify/:
    post:
      consumes:
      - application/json
      description: Verifies a one-time code sent by email.
      parameters:
      - description: '{email, otp, type}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Verify email OTP
      tags:
      - auth
  /api/auth/password/change/:
    post:
      consumes:
      - application/json
      description: Changes the current user's password.
      parameters:
      - description: '{old_password, new_password1, new_password2}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /api/auth/password/reset/:
    post:
      consumes:
      - application/json
      description: Emails a password-reset link. Always returns success to avoid leaking
        registered emails.
      parameters:
      - description: '{email}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Request password reset
      tags:
      - auth
  /api/auth/password/reset/confirm/:
    post:
      consumes:
      - application/json
      description: Resets the password using a uid + token from the reset email.
      parameters:
      - description: '{uid, token, new_password1, new_password2}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      summary: Confirm password reset
      tags:
      - auth
...
  /api/user/email-change/:
    post:
      consumes:
      - application/json
      description: Emails a verification link to the new address.
      parameters:
      - description: '{newEmail}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Request email change
      tags:
      - auth
  /api/user/email-change/confirm/:
    post:
      consumes:
      - application/json
      description: Applies a pending email change using the emailed token.
      parameters:
      - description: '{token}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.Error'
      security:
      - BearerAuth: []
      summary: Confirm email change
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"